[PULB-4] :recycle: 배치 설정에 대한 주석 작성
